+++ index.js
@@ -4,6 +4,7 @@
  * These are ESLint rules that have corresponding and recommended Biome rules.
  */
 module.exports = {
+  extends: ["./eslint-config-prettier.js"],
   rules: {
     "constructor-super": "off",
     "default-case-last": "off",