+++ scripts/index.test.ts
@@ -4,8 +4,10 @@ import path from "path"
 import { spawnSync } from "bun"
 
 const rootPath = path.resolve(import.meta.dir, "..")
-const indexJsPath = path.resolve(rootPath, "index.js")
-const indexJsContent = fs.readFileSync(indexJsPath, "utf-8")
+const indexPath = path.resolve(rootPath, "index.js")
+const prettierPath = path.resolve(rootPath, "eslint-config-prettier.js")
+const indexContent = fs.readFileSync(indexPath, "utf-8")
+const prettierContent = fs.readFileSync(prettierPath, "utf-8")
 
 test("index.js is a valid file and can be used by eslint", () => {
   // In the root there is a .eslintrc that uses the index.js in the extends.
@@ -26,7 +28,7 @@ test("index.js has rules from different plugins and includes extra rules", () =>
     "simple-import-sort/imports",
   ]
 
-  expect(rulesToCheck.every((rule) => indexJsContent.includes(rule))).toBeTrue()
+  expect(rulesToCheck.every((rule) => indexContent.includes(rule))).toBeTrue()
 })
 
 test("TS extensions should be added to index.js", () => {
@@ -35,5 +37,10 @@ test("TS extensions should be added to index.js", () => {
     "@typescript-eslint/default-param-last",
   ]
 
-  expect(rulesToCheck.every((rule) => indexJsContent.includes(rule))).toBeTrue()
+  expect(rulesToCheck.every((rule) => indexContent.includes(rule))).toBeTrue()
+})
+
+test("eslint-config-prettier is used and is valid", () => {
+  expect(indexContent).toContain('extends: ["./eslint-config-prettier.js"],')
+  expect(prettierContent).toContain('"react/jsx-indent": "off"')
 })